fixing login useEffect

+++ backend/validate/models.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 from typing import Optional
+from sqlalchemy import BigInteger
 
 class IdCheckRequest(BaseModel):
     id: int
@@ -13,7 +14,7 @@ Base = declarative_base()
 class Id(Base):
     __tablename__ = "ids"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
 
     def __repr__(self):
         return f"Id(id={self.id})"
